apps/api/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
celery>=5.4,<5.5
redis>=5.0,<5.1
boto3>=1.35,<1.36
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<1.1
Pillow>=11.0,<11.1
